A couple of changes to the Java interface:   - Variable's id are of long type (like space dimension arguments);   - added NOT_EQUAL value to enumeration Relation_Symbol.

+++ interfaces/Java/parma_polyhedra_library/Variable_Stringifier.java
@@ -29,5 +29,5 @@ package parma_polyhedra_library;
   method \c setStringifier.
 */
 public interface Variable_Stringifier {
-    public String stringify(int var_id);
+    public String stringify(long var_id);
 }
